refactor(be): move all overview related functions in model layer to each file
