+++ backend/src/services/dashboard/overview/topStayed.ts
@@ -1,4 +1,4 @@
-import { findTopStayed } from "@/models/dashboard/overview";
+import { findTopStayed } from "@/models/dashboard/overview/topStayed";
 
 export const retrieveTopStayed = async (limit: number) => {
   const retrievedTopStayed = await findTopStayed(limit);
