+++ backend/src/services/dashboard/overview/visitors.ts
@@ -1,5 +1,5 @@
 import { VisitorsQueryParamsDTO } from "@/types/DTO/dashboard/overview";
-import { findVisitors } from "@/models/dashboard/overview";
+import { findVisitors } from "@/models/dashboard/overview/visitors";
 
 export const retrieveVisitors = ({
   startDate,
